api/requirements.txt
requests==2.28.2
python-dotenv==0.21.1
pyyaml==6.0
requests-oauthlib==1.3.1 
munch==2.5.0
flask==2.2.3
flask-cors==3.0.10
flask-restful==0.3.9